Improve the way BpeTrainer creates a BPE
Before these modifications, the BPE constructed by the BpeTrainer didn't
account for some of its configuration. So if the Trainer was using some
suffix or prefix for example, it didn't instantiate the produced BPE
with these options.
Also any special tokens that are used during training, should be
automatically added to the Tokenizer with the Model. This is now handled
properly.