Rust - Fix offsets when merging multiple sequences
When the input sequence gets split into multiple sub-sequences,
there may be changes in the offsets (original <=> normalized) that
don't get reverted when merging back to one single sequence.
So in order to avoid this, we have to convert back to original offsets
before actually merging the various encodings and normalized strings back
together.